+++ R/goesaodc_createScanPoints.R
@@ -3,21 +3,9 @@
 #' @title Create AOD points from GOES scans
 #' 
 #' @description Creates a \code{SpatialPointsDataFrame} of AOD readings from a 
-#' GOES scan, or a list of \code{SpatialPointsDataFrame}s from a series of 
-#' scans.
+#' GOES scan file, or a list of \code{SpatialPointsDataFrame}s from a series of 
+#' files.
 #' 
-#' @details A single scan can be identified by either giving its 
-#' \code{filename}, or by providing the \code{satID}, \code{datetime}, and 
-#' \code{timezone} (which will be used to determine the closest matching scan). 
-#' A series of scans cannot be specified using filenames. The satellite + time 
-#' info must be given along with an \code{endtime} so that all scans starting at
-#' \code{datetime} up to (but not including) \code{endtime} will be processed.
-#' 
-#' @param satID ID of the source GOES satellite ('G16' or 'G17').
-#' @param datetime Datetime in Ymd HMS format or a \code{POSIXct}.
-#' @param endtime End time in Ymd HMS format or a \code{POSIXct}.
-#' @param timezone Timezone used to interpret \code{datetime} and 
-#' \code{endtime}; Defaults to UTC.
 #' @param filename Name of a scan file.
 #' @param bbox Bounding box for the region of interest. All points outside of 
 #' this area will be removed. Defaults to CONUS.
@@ -32,31 +20,27 @@
 #' 
 #' bboxOregon <- c(-125, -116, 42, 47)
 #' 
-#' # Create points for a scan specified by satellite + time
-#' goesaodc_createScanPoints(
+#' scanFiles <- goesaodc_listScanFiles(
 #'   satID = "G17",
-#'   datetime = "2020-09-08 12:30",
-#'   timezone = "America/Los_Angeles",
-#'   bbox = bboxOregon,
+#'   datetime = "2020-09-08 12:00",
+#'   endtime = "2020-09-08 13:00",
+#'   timezone = "America/Los_Angeles"
 #' )
 #' 
-#' # Create points for a scan file
+#' # Create an SPDF from a single scan file
 #' goesaodc_createScanPoints(
-#'   filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc",
+#'   filename = scanFiles[1],
 #'   bbox = bboxOregon,
 #'   dqfLevel = 2
 #' )
 #' 
-#' # Create points for multiple scans over a time range
+#' # Create a list of SPDFs from multiple scan files
 #' goesaodc_createScanPoints(
-#'   satID = "G17",
-#'   datetime = "2020-09-08 12:00",
-#'   endtime = "2020-09-08 13:00",
-#'   timezone = "America/Los_Angeles",
+#'   filename = scanFiles,
 #'   bbox = bboxOregon
 #' )
 #' 
-#' # Create points for a faulty scan
+#' # Create an empty SPDF from a faulty scan
 #' goesaodc_createScanPoints(
 #'   filename = "OR_ABI-L2-AODC-M6_G17_s20202522231174_e20202522233547_c20202522235327.nc",
 #'   bbox = bboxOregon
@@ -64,10 +48,6 @@
 #' }
 
 goesaodc_createScanPoints <- function(
-  satID = NULL,
-  datetime = NULL,
-  endtime = NULL,
-  timezone = "UTC",
   filename = NULL,
   bbox = bbox_CONUS,
   dqfLevel = 3
@@ -75,26 +55,7 @@ goesaodc_createScanPoints <- function(
   
   # ----- Validate parameters --------------------------------------------------
   
-  if ( is.null(filename) ) {
-    
-    MazamaCoreUtils::stopIfNull(satID)
-    
-    satID <- toupper(satID)
-    if ( !(satID %in% c("G16", "G17")) )
-      stop("Parameter 'satID' must be either 'G16' or 'G17'")
-    
-    MazamaCoreUtils::stopIfNull(datetime)
-    
-    # Use timezone from POSIXct datetime, or the given timezone
-    if ( lubridate::is.POSIXt(datetime) ) {
-      timezone <- lubridate::tz(datetime)
-    } else {
-      MazamaCoreUtils::stopIfNull(timezone)
-      if ( !timezone %in% OlsonNames() )
-        stop(sprintf("timezone \"%s\" is not recognized", timezone))
-    }
-    
-  }
+  MazamaCoreUtils::stopIfNull(filename)
   
   if ( !(dqfLevel %in% c(0, 1, 2, 3)) )
     stop(paste0("Parameter 'dqfLevel' must be 0, 1, 2, or 3"))
@@ -103,17 +64,22 @@ goesaodc_createScanPoints <- function(
   
   result <- try({
     
-    if ( is.null(filename) && !is.null(endtime) ) {
+    if ( length(filename) == 1 ) {
       
-      # Create a multi-scan spdf list
-      filenames <- goesaodc_listScanFiles(
-        satID = satID,
-        datetime = datetime,
-        endtime = endtime,
-        timezone = timezone,
-        useRemote = TRUE
+      # Create a single SPDF
+      spdf <- goesaodc_createSingleScanPoints(
+        filename = filename,
+        bbox = bbox,
+        dqfLevel = dqfLevel
       )
       
+      return(spdf)
+      
+    } else {
+      
+      filenames <- filename
+      
+      # Create a list of SPDFs
       spdfList <- list()
       for ( filename in filenames ) {
         
@@ -134,19 +100,6 @@ goesaodc_createScanPoints <- function(
       
       return(spdfList)
       
-    } else {
-      
-      spdf <- goesaodc_createSingleScanPoints(
-        satID = satID,
-        datetime = datetime,
-        timezone = timezone,
-        filename = filename,
-        bbox = bbox,
-        dqfLevel = dqfLevel
-      )
-      
-      return(spdf)
-      
     }
     
   }, silent = TRUE)
@@ -155,7 +108,7 @@ goesaodc_createScanPoints <- function(
   # AOD and DQF values for the requested satellite
   if ( "try-error" %in% class(result) ) {
     warning(result, immediate. = TRUE)
-    spdf <- goesaodc_createEmptyScanPoints(satID, filename, bbox)
+    spdf <- goesaodc_createEmptyScanPoints(filename = filename, bbox = bbox)
     return(spdf)
   }
   
@@ -164,16 +117,8 @@ goesaodc_createScanPoints <- function(
 
 #' @title Create AOD points from a single GOES scan
 #' 
-#' @description Creates a \code{SpatialPointsDataFrame} from a GOES scan.
+#' @description Creates a \code{SpatialPointsDataFrame} from a GOES scan file.
 #' 
-#' @details A scan can be identified by either giving its \code{filename}, or by
-#' providing the \code{satID}, \code{datetime}, and \code{timezone} (which will 
-#' be used to determine the closest matching scan). 
-#' 
-#' @param satID ID of the source GOES satellite ('G16' or 'G17').
-#' @param datetime Datetime in Ymd HMS format or a \code{POSIXct}.
-#' @param timezone Timezone used to interpret \code{datetime} and 
-#' \code{endtime}; Defaults to UTC.
 #' @param filename The name of the scan file.
 #' @param bbox Bounding box for the region of interest; Defaults to CONUS.
 #' @param dqfLevel Allowed data quality level. All readings with a DQF value
@@ -187,26 +132,19 @@ goesaodc_createScanPoints <- function(
 #' 
 #' bboxOregon <- c(-125, -116, 42, 47)
 #' 
-#' # Create points from a scan specified by satellite and time
-#' MazamaSatelliteUtils:::goesaodc_createSingleScanPoints(
+#' scanFile <- goesaodc_listScanFiles(
 #'   satID = "G17",
-#'   datetime = "2020-09-08 17:30",
-#'   timezone = "America/Los_Angeles",
-#'   bbox = bboxOregon
+#'   datetime = "2020-09-08 12:30",
+#'   timezone = "America/Los_Angeles"
 #' )
 #' 
-#' # Create points from a scan file
 #' MazamaSatelliteUtils:::goesaodc_createSingleScanPoints(
-#'   filename = "OR_ABI-L2-AODC-M6_G17_s20202530031174_e20202530033547_c20202530035523.nc",
-#'   bbox = bboxOregon,
-#'   dqfLevel = 2
+#'   filename = scanFile,
+#'   bbox = bboxOregon
 #' )
 #' }
 
 goesaodc_createSingleScanPoints <- function(
-  satID = NULL,
-  datetime = NULL,
-  timezone = "UTC",
   filename = NULL,
   bbox = bbox_CONUS,
   dqfLevel = 3
@@ -214,72 +152,17 @@ goesaodc_createSingleScanPoints <- function(
   
   # ----- Validate parameters --------------------------------------------------
   
-  if ( is.null(filename) ) {
-    
-    MazamaCoreUtils::stopIfNull(satID)
-    
-    satID <- toupper(satID)
-    if ( !(satID %in% c("G16", "G17")) )
-      stop("Parameter 'satID' must be either 'G16' or 'G17'")
-    
-    MazamaCoreUtils::stopIfNull(datetime)
-    
-    # Use timezone from POSIXct datetime, or passed in timezone
-    if ( lubridate::is.POSIXt(datetime) ) {
-      timezone <- lubridate::tz(datetime)
-    } else {
-      if ( !timezone %in% OlsonNames() ) {
-        stop(sprintf("timezone \"%s\" is not recognized", timezone))
-      }
-    }
-    
-  }
+  MazamaCoreUtils::stopIfNull(filename)
   
   if ( !(dqfLevel %in% c(0, 1, 2, 3)) )
     stop(paste0("Parameter 'dqfLevel' must be 0, 1, 2, or 3"))
   
-  # ----- Get scan file --------------------------------------------------------
-  
-  if ( is.null(filename) ) {
-    
-    # Find which scan file is closest to the requested time
-    filename <- goesaodc_listScanFiles(
-      satID = satID,
-      datetime = datetime,
-      timezone = timezone,
-      useRemote = TRUE
-    )
-    
-  }
+  # ----- Create spatial points ------------------------------------------------
   
   # Download the file if it isn't available locally
   if ( !(filename %in% list.files(getSatelliteDataDir())) ) {
-    
-    if ( is.null(satID) ) {
-      MazamaCoreUtils::stopIfNull(filename)
-      filePattern <- "OR_ABI-L2-AODC-M[0-9]_(G16|G17)_s[0-9]+_e[0-9]+_c[0-9]+\\.nc"
-      satID <- stringr::str_match(filename, filePattern)[1,2]
-    }
-    
-    fileUrl <- paste0(
-      "https://tools-1.airfire.org/Satellite/",
-      ifelse(satID == "G16", "GOES-16/AODC/", "GOES-17/AODC/"),
-      filename
-    )
-    
-    filePath <- file.path(getSatelliteDataDir(), filename)
-    
-    utils::download.file(
-      fileUrl,
-      destfile = filePath, 
-      quiet = TRUE, 
-      method = "auto", 
-      mode = "wb"
-    )
-    
+    goesaodc_downloadScanFiles(filenames = filename)
   }
-  
-  # ----- Create spatial points ------------------------------------------------
   
   nc <- goesaodc_openScanFile(filename)